reorder zigbee_nwk_commands into modules
